service1/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests>=2.28.0
python-dotenv==1.0.0
pymysql
